Project 4/requirements.txt
aniso8601==8.0.0
certifi==2019.9.11
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
enum34==1.1.10
Flask==1.1.1
Flask-RESTful==0.3.7
idna==2.8
ipaddress==1.0.23
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.9
PyJWT==1.7.1
pytz==2020.4
requests==2.22.0
six==1.15.0
urllib3==1.25.11
Werkzeug==1.0.1
